Internal: Error handling in new Onboarding [ED-23026]
