+++ app/modules/e-onboarding/storage/onboarding-progress-manager.php
@@ -14,7 +14,6 @@ class Onboarding_Progress_Manager {
 	const PROGRESS_OPTION_KEY = 'elementor_e_onboarding_progress';
 	const CHOICES_OPTION_KEY = 'elementor_e_onboarding_choices';
 	const DEFAULT_TOTAL_STEPS = 5;
-	const ALLOWED_THEMES = [ 'hello-elementor', 'hello-biz' ];
 
 	private static ?Onboarding_Progress_Manager $instance = null;
 
@@ -137,44 +136,6 @@ class Onboarding_Progress_Manager {
 		}
 
 		return $this->save_choices( $choices );
-	}
-
-	public function install_and_activate_theme( string $theme_slug ): bool {
-		if ( ! in_array( $theme_slug, self::ALLOWED_THEMES, true ) ) {
-			return false;
-		}
-
-		if ( ! current_user_can( 'install_themes' ) || ! current_user_can( 'switch_themes' ) ) {
-			return false;
-		}
-
-		$theme = wp_get_theme( $theme_slug );
-
-		if ( ! $theme->exists() ) {
-			if ( ! function_exists( 'request_filesystem_credentials' ) ) {
-				require_once ABSPATH . 'wp-admin/includes/file.php';
-			}
-
-			if ( ! class_exists( '\Theme_Upgrader' ) ) {
-				require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
-			}
-
-			if ( ! class_exists( '\WP_Ajax_Upgrader_Skin' ) ) {
-				require_once ABSPATH . 'wp-admin/includes/class-wp-ajax-upgrader-skin.php';
-			}
-
-			$skin = new \WP_Ajax_Upgrader_Skin();
-			$upgrader = new \Theme_Upgrader( $skin );
-			$result = $upgrader->install( "https://downloads.wordpress.org/theme/{$theme_slug}.latest-stable.zip" );
-
-			if ( is_wp_error( $result ) || ! $result ) {
-				return false;
-			}
-		}
-
-		switch_theme( $theme_slug );
-
-		return true;
 	}
 
 	public function reset(): void {
